+++ noodles-cram/CHANGELOG.md
@@ -6,6 +6,8 @@
 
   * cram/reader: Implement `sam::AlignmentReader`.
 
+  * cram/record/features: Add conversion from SAM CIGAR operations.
+
 ### Changed
 
   * cram/record: Move `mapping_quality`, `read_name`, and `quality_scores` to